+++ src/index.ts
@@ -2,3 +2,6 @@ import 'reflect-metadata';
 
 export * from "./decorators/ApiResponseDecorator";
 export * from "./classes/Error";
+export * from "./functions/createErrorResponse";
+export * from "./functions/createSuccessResponse";
+export * from "./types/ResponseDto";