+++ src/types/Options.ts
@@ -1,0 +1,7 @@
+import { SuccessStatus } from "./SuccessStatus";
+
+export type Options = {
+  onSuccess?: { status?: SuccessStatus, message?: string };
+  onError?: { status?: number, message?: string };
+  enableDebug?: boolean,
+};